reworked modules loading mechanics
