+++ modules/pastebin_scraper.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python3
+# https://github.com/empty-circle/empty-circle.git
+# The Empty Circle Reconnaissance Framework - 2023
+
+
+# Wrapper function for Clearweb Scraper
+def clearweb_wrapper():
+    call_process(f"python3 scrappy_pup.py")
