+++ modules/scanhub.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python3
+# https://github.com/empty-circle/empty-circle.git
+# The Empty Circle Reconnaissance Framework - 2023
+
+
+# Wrapper function for Script Scanning Hub
+def script_scan_wrapper():
+    call_process(f"sudo sh ec_scanhub.sh")
