+++ modules/scrappy_pup.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python3
+# https://github.com/empty-circle/empty-circle.git
+# The Empty Circle Reconnaissance Framework - 2023
+
+
+# Wrapper function for Pastebin Crawler
+def pastebin_crawler_wrapper():
+    call_process(f"python3 pastebin_scraper.py")
